reduce example to test simple case of one mutation in a command with a promise
